Add scripts/ directory to path, relative to text2vid.py file
Fixes extension failing to load when installing the extension with an unusual directory name (like mine).

Copies what is done in api_t2v.py

I think this accounts for the other additions to the path as well, but I didn't want to touch those in case something broke on a different setup.